Added example for Cocoa Collections

+++ objc/JsonLiteCollections.m
@@ -124,7 +124,7 @@ typedef struct JsonLiteDictionaryBucket {
     if (index >= count) {
         id cls = NSStringFromClass([self class]);
         id sel = NSStringFromSelector(_cmd);
-        NSString *str = [NSString stringWithFormat:@"*** -[%@ %@]: index (%d) beyond bounds (%d)", cls, sel, index, count];
+        NSString *str = [NSString stringWithFormat:@"*** -[%@ %@]: index (%d) beyond bounds (%d)", cls, sel, (int)index,(int)count];
         exc = [NSException exceptionWithName:NSRangeException
                                       reason:str
                                     userInfo:nil];
